+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This changelog also contains an important changes in dependencies.
 
 ### Fixed
 - (cairo-backend) Text layout.
+- (cairo-backend) Rendering of a zero length subpath with a square cap.
 - (resvg) Recursive SVG images via `image` tag.
 - (resvg) Bbox calculation of the text with rotate.
 - (qt-api) SVG from QByteArray loading when data is invalid.